requirements.txt
asgiref==3.5.2
attrs==22.1.0
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coloredlogs==15.0.1
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
Django==4.1
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
filelock==3.8.0
html5lib==1.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.4
jmespath==1.0.1
lxml==4.9.1
Markdown==3.4.1
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyreadline3==3.4.1
python-scrapyd-api==2.1.2
pytz==2022.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
scrapy-djangoitem==1.1.1
scrapyd==1.3.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tldextract==3.3.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
w3lib==2.0.1
webencodings==0.5.1
zipp==3.8.1
zope.interface==5.4.0
